Created New DLL for Extendability Control. Added new DevEnvXXXSvc functions.

+++ NasuTek.DevEnvironment.GPLComponents/Documents/TextEditor.cs
@@ -1,5 +1,4 @@
 ﻿using System.IO;
-using NasuTek.DevEnvironment.Extensibility.Workbench;
 using System;
 using System.Collections.Generic;
 using System.ComponentModel;
@@ -8,11 +7,14 @@ using System.Drawing;
 using System.Linq;
 using System.Text;
 using System.Windows.Forms;
+using NasuTek.DevEnvironment.Extensibility.Workbench;
+using NasuTek.DevEnvironment.Extensibility;
 
 namespace NasuTek.DevEnvironment.Documents
 {
     public partial class TextEditor : DevEnvDocument {
         private string m_Path;
+        private bool isFile;
 
         public TextEditor()
         {
@@ -20,19 +22,28 @@ namespace NasuTek.DevEnvironment.Documents
         }
 
         public override void Open(DocumentMetadata openObj) {
-            Text = Path.GetFileName(openObj.FilePath);
-            textEditorControl1.Text = File.ReadAllText(openObj.FilePath);
-            m_Path = openObj.FilePath;
+            if (openObj.IsFile)
+            {
+                isFile = true;
+
+                Text = Path.GetFileName(openObj.FilePath);
+                textEditorControl1.Text = File.ReadAllText(openObj.FilePath);
+                m_Path = openObj.FilePath;
+            }
         }
 
         public override void Save() {
-            File.WriteAllText(m_Path, textEditorControl1.Text);
+            if (isFile)
+                File.WriteAllText(m_Path, textEditorControl1.Text);
         }
 
         public override void SaveAs(string path) {
-            m_Path = path;
-            Text = Path.GetFileName(m_Path);
-            Save();
+            if (isFile)
+            {
+                m_Path = path;
+                Text = Path.GetFileName(m_Path);
+                Save();
+            }
         }
     }
 }
